Display configuration read from server, not from file.
When importing a new one, the file is immediately sent to the server and
inserted to the DB.

Changing phase colors is immediately forwarded to the DB without the
need of any confirmation.

When the model state is MODELLING (diagram is not empty), changing phase
colors and creating task types is still possible.