Changed scope type NONE to UNIT.
This type is a shortcut to refer to all attributes, hence a scope at the
unit level.

For dependencies, this type cannot be used if the source and target
tasks have different phases, because there only a subset of attributes
can be used, so selecting results in a scope bigger than the unit scope
of either phase.