Improved logic for interdependency between scope type and attributes.
The entity which references the scope must know the total attributes the
scope refers to.
Then depending on the type, the following rules apply:

If the type is UNIT, all attributes are set.
If the type is GLOBAL, no attributes are set.

If the type is ATTRIBUTES, it might change to UNIT or GLOBAL depending
on whether all or no attributes are set, respectively.