Implemented tentative import of diagram network + minor improvements.
A diagram can now be imported into an existing project, overwriting the
exiting diagram (task + dependencies), but using the existing
configuration.

Matching the new diagram and the existing configuration is done by name.
If no match is found:
- For attributes, phases, and crafts, null is used
- For task definitions, a new task definition is added
  (so the resulting diagram has all the names)

In particular, now there can be task definitions without phases.
The user can select the phase of a task definition in the UI if it does
not have any phase yet, but then that phase is fixed.

For matching locations, the tree is navigated from the root along the
path described by the location.
If the path turns out to be invalid, the most recent valid node is used,
or the global location if there is none.

Scopes and orderings are not well considered yet.

More improvements:
- BUGFIX: creating dependencies did not work properly when the target is
  moved slightly.
- When whole models are imported, eventual errors are shown in a
  table-dialog with message, property path, and invalid value for each
  error.
- Implemented a check whether the scopes of dependencies are subsets of
  the attributes of the phases of source and target task. Previously,
  this was only considered for providing the user only valid attributes
  to be selected.
- In the diagram, task properties are now grouped into sections, each of
  which can be collapsed and expanded.
- Also, the 'Done' and 'Cancel' buttons in the property views are now
  fixed at the bottom.
- Added more flexibility to CachingTranslators for supporting the use
  case of importing diagrams.