Apply graph partitioning recursively within AcyclicOrientationFinder.
Use parallel streams to scale the algorithm with CPU cores.

Extracted resolving edges from AcyclicOrientationFinder as a new service
"DisjunctiveEdgeResolver"

Also, an OrientationResult (formerly AcyclicOrientationFinder.Result)
is now represented as a tree of results corresponding to the graph
partitions.
From a successful OrientationResult, an actual acyclic orientation can
be constructed on demand, whereas it can be omitted if not needed, e.g.
when checking consistency.

The GraphController now uses this to also expose an orientation of a
model if one exists.
Moreover, graphs (both disjunctive and oriented) are now exported in
JSON).

When adding an arc or edge to a DisjunctiveGraph where some nodes are
not contained in the graph, the arc or edge is ignored.