Renamed "Task Type" / "Task Definition" to "Activity".
Also simplified filenames and module names of several angular
components.

BUGFIX: simple search queries only worked if tasks have both a phase and
a craft (left outer join was required)

BUGFIX: Crafts showed a description, but DB and server did not know of
it.

Attributes now have an additional flag "perPhase", which can be used to
indicate that if it is used in several phases, it has different meanings
and should not be considered shared. This means that it cannot be used
in the scope of a dependency between tasks of different phases. Useful
e.g. to distinguish "facade units" from "room units".
If there are multiple refer to room units, then one needs to define two
separate unit attributes, e.g. "f-unit" and "r-unit".